Simple Banking System (ATM simulator)
Simulate how a very simple ATM works. Accepts a credit card. Checks if it's numbers are valid and then if the credit card exists in the Bank database. Checks if the password is right. It allows the user to increase or decrease the account balance ("withdraw" or "deposit" money), check the account balance and some other simple options. It also creates new users and adds them to the database, which is hold on a SQL file. An empty SQL file will be automatically created on the same directory of this python file. The SQL file will be fill in with the new accounts that are created with this Bank Simulator.