Coffee machine simulator
It simulates the screen of a coffee machine. It gives 5 options:
buy, fill, take, remaining and exit
it has initial resourses to make coffee, and if you want to buy coffee, it will "make" the coffee only if it has enough resourses. To fill in the machine, select "fill". If you select "take", It will "give" you al the collected money from the already sold coffees.
"Remaining" shows you how much of the ingredients the machine has.